+++ x/evidence/keeper/querier_test.go
@@ -3,6 +3,8 @@ package keeper_test
 import (
 	"strings"
 
+	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
+
 	"github.com/cosmos/cosmos-sdk/std"
 
 	"github.com/cosmos/cosmos-sdk/x/evidence/exported"
@@ -18,7 +20,7 @@ const (
 func (suite *KeeperTestSuite) TestQueryEvidence_Existing() {
 	ctx := suite.ctx.WithIsCheckTx(false)
 	numEvidence := 100
-	cdc := std.NewAppCodec(suite.app.Codec())
+	cdc := std.NewAppCodec(suite.app.Codec(), codectypes.NewInterfaceRegistry())
 
 	evidence := suite.populateEvidence(ctx, numEvidence)
 	query := abci.RequestQuery{
@@ -37,7 +39,7 @@ func (suite *KeeperTestSuite) TestQueryEvidence_Existing() {
 
 func (suite *KeeperTestSuite) TestQueryEvidence_NonExisting() {
 	ctx := suite.ctx.WithIsCheckTx(false)
-	cdc := std.NewAppCodec(suite.app.Codec())
+	cdc := std.NewAppCodec(suite.app.Codec(), codectypes.NewInterfaceRegistry())
 	numEvidence := 100
 
 	suite.populateEvidence(ctx, numEvidence)
@@ -53,7 +55,7 @@ func (suite *KeeperTestSuite) TestQueryEvidence_NonExisting() {
 
 func (suite *KeeperTestSuite) TestQueryAllEvidence() {
 	ctx := suite.ctx.WithIsCheckTx(false)
-	cdc := std.NewAppCodec(suite.app.Codec())
+	cdc := std.NewAppCodec(suite.app.Codec(), codectypes.NewInterfaceRegistry())
 	numEvidence := 100
 
 	suite.populateEvidence(ctx, numEvidence)
@@ -73,7 +75,7 @@ func (suite *KeeperTestSuite) TestQueryAllEvidence() {
 
 func (suite *KeeperTestSuite) TestQueryAllEvidence_InvalidPagination() {
 	ctx := suite.ctx.WithIsCheckTx(false)
-	cdc := std.NewAppCodec(suite.app.Codec())
+	cdc := std.NewAppCodec(suite.app.Codec(), codectypes.NewInterfaceRegistry())
 	numEvidence := 100
 
 	suite.populateEvidence(ctx, numEvidence)